chenge wmma scale
